Make assumptions when other strategies don't work
If left with an unsolved grid, and elimination and inferring
does not change the state of the grid, pick a cell (The first
unsolved one) and make an assumption based on the possible
values it could hold.

If the grid is solvable: great! If it isn't: reset the grid to
before the assumption and try another value.